functionality with Managers
