+++ Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib/connectionManager.h
@@ -20,7 +20,7 @@ typedef struct {
 
    void print_communication(); 
    void initWiFi();
-   void connectMQTT(connectionManager);
+   bool connectMQTT(connectionManager);
    void connectWiFi(connectionManager);
    void reconnectWiFi(connectionManager);
    void resetWifi(connectionManager);
