+++ Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib/myCommon.h
@@ -3,7 +3,7 @@
 
 
     #include "receiverBoard.h"
-    #include "Communication.h"
+    //#include "Communication.h"
     #include "EnergyMonitoring.h"
     #include "app_config.h"
